refining structure
